Remove tenant information in the REST API url of attachments.
